Add support for locales
EUR and USD are supported now

+++ data.processor.php
@@ -6,35 +6,51 @@ if(isset($_POST['form_type']) AND !empty($_POST['form_type'])) :
 	
 		
 		case 'nwatch':
-	
-			// Locate the old password file 
-			$path    	= 'admin/';
-			$files 		= scandir($path);
 			
-			foreach($files as $file):
+			if(isset($_POST['nwatch_password']) AND !empty($_POST['nwatch_password'])) : 
 			
-				if($file != '.' AND $file != '..' AND $file != 'index.php'): 
-					
-					// Delete the file 
-					unlink($path.''.$file); 
-					
+				// Locate the old password file 
+				$path    	= 'admin/';
+				$files 		= scandir($path);
+				
+				foreach($files as $file):
+				
+					if($file != '.' AND $file != '..' AND $file != 'index.php'): 
+						
+						// Delete the file 
+						unlink($path.''.$file); 
+						
+					endif; 
+				
+				endforeach;
+				
+				// Create a random name for the password file 
+				
+				$length = rand(10,100);    
+				$file	= substr(str_shuffle('0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ'),1,$length);
+				
+				// Store the password 
+			
+				$pwd_file = fopen($path.''.$file, "w"); 
+				
+				$hash = password_hash($_POST['nwatch_password'], PASSWORD_DEFAULT);
+				
+				fwrite($pwd_file, $hash);
+				fclose($pwd_file);
+				
+			endif; 
+			
+			// locale
+			
+			if(isset($_POST['locale']) AND !empty($_POST['locale'])) : 
+				
+				if(isset($_COOKIE['nW_locale'])) : 
+					setcookie("nW_locale", '',  time()-3600 );
 				endif; 
+				
+				setcookie("nW_locale", $_POST['locale'],  time() + (10 * 365 * 24 * 60 * 60) );
 			
-			endforeach;
-			
-			// Create a random name for the password file 
-			
-			$length = rand(10,100);    
-			$file	= substr(str_shuffle('0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ'),1,$length);
-			
-			// Store the password 
-		
-			$pwd_file = fopen($path.''.$file, "w"); 
-			
-			$hash = password_hash($_POST['nwatch_password'], PASSWORD_DEFAULT);
-			
-			fwrite($pwd_file, $hash);
-			fclose($pwd_file);
+			endif; 
 		
 		break;
